split date and time setter, made types file, async
