+++ api/src/index.ts
@@ -2,7 +2,8 @@ import express,{Request,Response} from 'express';
 import bodyParser from 'body-parser';
 import cors from 'cors';
 import sqlite3 from 'sqlite3';
-import preDb from './preDb'; 
+import preDb from './preDb';
+import {deskState, deskType} from './OwnTypes' 
 
 const app = express();
 const port = 3001;
@@ -12,14 +13,7 @@ const corsOptions = {
   origin: 'http://localhost:3000'
 }
 
-const deskState = {
-  free: 'free',
-  checkedIn: 'checked in',
-  reserved: 'reserved',
-  unavailable: 'unavailable'
-} as const;
-
-let db:any;
+let db: any;
 app.use(express.static('public'));
 app.use(cors(corsOptions))
 app.use(bodyParser.urlencoded({extended: false}));
@@ -42,9 +36,9 @@ app.get("/api/desk/:deskId", async (req: Request, res: Response) => {
 });
 
 app.patch("/api/desk/:deskId", async (req: Request, res: Response) => {
-  const initialDeskState = await db.get(
-    "SELECT * FROM desk WHERE desk_id = (?)",req.params.deskId).desk_state;
-  const newDeskState = initialDeskState === deskState.free ? deskState.checkedIn : deskState.free;
+  const initialDesk = await db.get(
+    "SELECT * FROM desk WHERE desk_id = (?)",req.params.deskId);
+  const newDeskState = (initialDesk.desk_state === deskState.checkedIn) ? deskState.free : deskState.checkedIn;
   await db.get("UPDATE desk SET desk_state = (?) WHERE desk_id = (?)",newDeskState, req.params.deskId);
   const desk = await db.get(
     "SELECT * FROM desk WHERE desk_id = (?)",req.params.deskId);
